perf: use sqrtf() instead of sqrt() in stream_ave RMS output (#166)
Replace sqrt() with sqrtf() in the RMS output computation
(line 184) since the result is stored in a float array.
Avoids unnecessary double promotion.

Identified via static LTO build + objdump inspection.

+++ src/coremods/COREMOD_memory/stream_ave.c
@@ -181,7 +181,7 @@ static MILK_HOT errno_t fpsexec(
                  i < xysize; i++)
             {
                 imgoutrms->array.F[i] =
-                    sqrt(imdataarrayPOW[i])
+                    sqrtf(imdataarrayPOW[i])
                     / (streamave_cntindex);
             }
             processinfo_update_output_stream(
